refactor(dashboard-tsr): dedup data-table-header display options (#1392)
Extract ViewToggle + the twice-duplicated density/column-visibility
DropdownMenu into a shared DisplayOptionsDropdown, and collapse the three
inline filter-draft literals into a blankFilterDraft helper. No behavior
or markup change.

Part of epic #1392
